backend/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
python-docx==0.8.11
pdfplumber==0.9.0
sentence-transformers
huggingface-hub
numpy
pandas
openai==0.27.8